16_input_form_validation/requirements.txt
Flask
Flask-WTF
WTForms
email_validator
python-dotenv